requirements.txt
alembic==0.9.5
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.0
Flask-WTF==0.14.2
infinity==1.4
intervals==0.8.0
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.1.14
SQLAlchemy-Utils==0.32.17
validators==0.12.0
Werkzeug==0.12.2
WTForms==2.1
WTForms-Alchemy==0.16.5
WTForms-Components==0.10.3
